Add Edit-Profile & webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
-import Video, { formatHashtags } from "../models/Video";
+import Video from "../models/Video";
+import User from "../models/User";
 
 /*
 callback
@@ -19,18 +20,31 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   //const id = req.params.id; 와 똑같음.
   const { id } = req.params;
-  const video = await Video.findById(id);
-  if (video) {
-    return res.render("watch", { pageTitle: `Watching ${video.title}`, video });
+
+  //mongoose에 이 ID가 Usermodel에서 왔다고 알려주기만 하면됨
+  const video = await Video.findById(id).populate("owner");
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
   }
-  return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
+  return res.render("watch", {
+    pageTitle: `Watching ${video.title}`,
+    video,
+  });
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  console.log(video);
   return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
@@ -57,18 +71,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   //NEW video document
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
-      meta: {
-        views: 0,
-        rating: 0,
-      },
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", {
@@ -80,6 +99,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "VIDEO NOT FOUND" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
